+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Container/CompilerPasses/ConfigureSchemaNamespacingCompilerPass.php
@@ -93,6 +93,7 @@ class ConfigureSchemaNamespacingCompilerPass extends AbstractConfigureSchemaName
             \PoPWPSchema\Pages\Module::class,
             \PoPWPSchema\Posts\Module::class,
             \PoPWPSchema\Blocks\Module::class,
+            \PoPWPSchema\Site\Module::class,
             \PoPWPSchema\TaxonomyMeta\Module::class,
             \PoPWPSchema\UserMeta\Module::class,
             \PoPWPSchema\Users\Module::class,