+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Module.php
@@ -58,6 +58,7 @@ class Module extends AbstractPluginModule
             \PoPWPSchema\Pages\Module::class,
             \PoPWPSchema\Posts\Module::class,
             \PoPWPSchema\Blocks\Module::class,
+            \PoPWPSchema\Site\Module::class,
             \PoPWPSchema\TaxonomyMeta\Module::class,
             \PoPWPSchema\UserMeta\Module::class,